Resolve bug related to how guid gets generated in StoreImageAnalysisTypePeople helper method.
